replace coaddition.py ; move back simple Spectrum object to frame.py ; working version of spectroperf
